+++ opentargets/__init__.py
@@ -10,10 +10,11 @@ logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
 class OpenTargetsClient(object):
-    '''
-    main class to use to get data from the Open Targets REST API available at targetvalidation.org, or your private instance
+    """
+    Main class to use to get data from the Open Targets REST API available at targetvalidation.org (or your
+    private instance)
 
-    '''
+    """
 
     _search_endpoint = '/public/search'
     _filter_associations_endpoint = '/public/association/filter'
@@ -26,11 +27,11 @@ class OpenTargetsClient(object):
     def __init__(self,
                  **kwargs
                  ):
-        '''
-
-        :param kwargs: all params forwarded to :class:`opentargets.conn.Connection` object
-
-        '''
+        """
+        Init the client and start a connection
+        Keyword Args:
+            **kwargs: all params forwarded to opentargets.conn.Connection object
+        """
         self.conn = Connection(**kwargs)
 
     def __enter__(self):
@@ -40,27 +41,66 @@ class OpenTargetsClient(object):
         self.close()
 
     def close(self):
-        # self.conn.close()
-        pass
+        self.conn.close()
 
     def search(self, query,**kwargs):
+        """
+        Search a string and return a list of objects form the search method of the REST API.
+        E.g. A returned object could be a target or a disease
+        Args:
+            query (str): string to search for
+        Keyword Args:
+            **kwargs are passed as other parameters to the /public/search method of the REST API
+
+        Returns:
+            IterableResult: Result of the query
+        """
         kwargs['q']=query
         result = IterableResult(self.conn)
         result(self._search_endpoint,**kwargs)
         return result
 
     def get_association(self,association_id, **kwargs):
+        """
+        Retrieve a specific Association object from the REST API provided its ID
+
+        Args:
+            association_id (str): Association ID
+        Keyword Args:
+            **kwargs are passed as other parameters to the /public/association method of the REST API
+
+        Returns:
+             IterableResult: Result of the query
+        """
         kwargs['id']= association_id
         result = IterableResult(self.conn)
         result(self._get_associations_endpoint, **kwargs)
         return result
 
     def filter_associations(self,**kwargs):
+        """
+        Retrieve a set of associations by applying a set of filters
+        Keyword Args:
+            **kwargs are passed as parameters to the /public/association/filterby method of the REST API
+
+        Returns:
+            IterableResult: Result of the query
+        """
         result = IterableResult(self.conn)
         result(self._filter_associations_endpoint, **kwargs)
         return result
 
     def get_associations_for_target(self, target, **kwargs):
+        """
+        Same as ``OpenTargetsClient.filter_associations`` but accept any string as `target` parameter and fires a search
+        if it is not an Ensembl Gene identifier
+        Args:
+            target (str): an Ensembl Gene identifier or a string to search for a gene mapping
+        Keyword Args:
+            **kwargs are passed as parameters to the /public/association/filterby method of the REST API
+        Returns:
+            IterableResult: Result of the query
+        """
         if not isinstance(target, str):
             raise AttributeError('target must be of type str')
         if not target.startswith('ENSG'):
@@ -74,6 +114,16 @@ class OpenTargetsClient(object):
         return self.filter_associations(target=target_id,**kwargs)
 
     def get_associations_for_disease(self, disease, **kwargs):
+        """
+        Same as ``OpenTargetsClient.filter_associations`` but accept any string as `disease` parameter and
+        fires a search if it is not a valid disease identifier
+        Args:
+            disease (str): a disease identifier or a string to search for a disease mapping
+        Keyword Args:
+            **kwargs are passed as parameters to the /public/association/filterby method of the REST API
+        Returns:
+            IterableResult: Result of the query
+        """
         if not isinstance(disease, str):
             raise AttributeError('disease must be of type str')
         results = self.filter_associations(disease=disease)
@@ -83,21 +133,49 @@ class OpenTargetsClient(object):
                 raise AttributeError('cannot find an disease id for disease {}'.format(disease))
             disease_id = search_result['id']
             logger.debug('{} resolved to id {}'.format(disease, disease_id))
-            results = self.filter_associations(disease=disease_id)
+            results = self.filter_associations(disease=disease_id, **kwargs)
         return results
 
     def get_evidence(self, evidence_id, **kwargs):
+        """
+        Retrieve a specific Evidence object from the REST API provided its ID
+        Args:
+            evidence_id:
+        Keyword Args:
+            **kwargs are passed as other parameters to the /public/evidence method of the REST API
+
+        Returns:
+             IterableResult: Result of the query
+        """
         kwargs['id']= evidence_id
         result = IterableResult(self.conn)
         result(self._get_evidence_endpoint, **kwargs)
         return result
 
     def filter_evidence(self,**kwargs):
+        """
+        Retrieve a set of evidence by applying a set of filters
+        Keyword Args:
+            **kwargs are passed as parameters to the /public/evidence/filterby method of the REST API
+
+        Returns:
+            IterableResult: Result of the query
+        """
         result = IterableResult(self.conn)
         result(self._filter_evidence_endpoint, **kwargs)
         return result
 
     def get_evidence_for_target(self, target, **kwargs):
+        """
+        Same as ``OpenTargetsClient.filter_evidence`` but accept any string as `target` parameter and fires a search
+        if it is not an Ensembl Gene identifier
+        Args:
+            target (str): an Ensembl Gene identifier or a string to search for a gene mapping
+        Keyword Args:
+            **kwargs are passed as parameters to the /public/evidence/filterby method of the REST API
+        Returns:
+            IterableResult: Result of the query
+        """
         if not isinstance(target, str):
             raise AttributeError('target must be of type str')
         if not target.startswith('ENSG'):
@@ -111,9 +189,19 @@ class OpenTargetsClient(object):
         return self.filter_evidence(target=target_id,**kwargs)
 
     def get_evidence_for_disease(self, disease, **kwargs):
+        """
+        Same as ``OpenTargetsClient.filter_evidence`` but accept any string as `disease` parameter and
+        fires a search if it is not a valid disease identifier
+        Args:
+            t (str): a disease identifier or a string to search for a disease mapping
+        Keyword Args:
+            **kwargs are passed as parameters to the /public/evidence/filterby method of the REST API
+        Returns:
+            IterableResult: Result of the query
+        """
         if not isinstance(disease, str):
             raise AttributeError('disease must be of type str')
-        results = self.filter_evidence(disease=disease)
+        results = self.filter_evidence(disease=disease, **kwargs)
         if not results:
             search_result = next(self.search(disease, size=1, filter='disease'))
             if not search_result:
@@ -124,6 +212,11 @@ class OpenTargetsClient(object):
         return results
 
     def get_stats(self):
+        """
+        Returns statistics about the data served by the REST API
+        Returns:
+            IterableResult: Result of the query
+        """
         result = IterableResult(self.conn)
         result(self._stats_endpoint)
         return result